# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kreatorboard API - Creator Operating System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /integrations/{provider}/connect:
    post:
      operationId: initiateIntegrationConnect
      tags: [integrations]
      summary: Start OAuth connection flow for a provider
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [gmail, google_calendar, discord, telegram, twitter, cal_com, wallet, website]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectIntegrationRequest"
      responses:
        "200":
          description: Connection initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectResponse"
  /integrations/{provider}/disconnect:
    post:
      operationId: disconnectIntegration
      tags: [integrations]
      summary: Disconnect an integration
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /integrations/{provider}/sync:
    post:
      operationId: triggerIntegrationSync
      tags: [integrations]
      summary: Trigger manual sync for a provider
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sync triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncJobResponse"
  /integrations/wallet/connect:
    post:
      operationId: connectWallet
      tags: [integrations]
      summary: Connect and verify a wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectWalletRequest"
      responses:
        "200":
          description: Wallet connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletConnection"
...
  /integrations/oauth/callback:
    get:
      operationId: oauthCallback
      tags: [integrations]
      summary: OAuth callback handler
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect after OAuth
...
  /inbox:
    get:
      operationId: getInbox
      tags: [inbox]
      summary: Get collaboration inbox
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, in_progress, accepted, rejected, archived]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Inbox items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCollaborations"
  /inbox/{id}:
    get:
      operationId: getCollaboration
      tags: [inbox]
      summary: Get single collaboration request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Collaboration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollaborationDetail"
    patch:
      operationId: updateCollaboration
      tags: [inbox]
      summary: Update collaboration status or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCollaborationRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollaborationDetail"
  /inbox/request:
    post:
      operationId: submitCollaborationRequest
      tags: [inbox]
      summary: Submit a collaboration request (from founder to creator)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitCollaborationRequest"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollaborationDetail"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    ConnectIntegrationRequest:
      type: object
      properties:
        redirectUrl:
          type: string
        metadata:
          type: object
          additionalProperties: true
    ConnectResponse:
      type: object
      properties:
        authUrl:
          type: string
          nullable: true
        message:
          type: string
        requiresRedirect:
          type: boolean
      required:
        - message
        - requiresRedirect
    ConnectWalletRequest:
      type: object
      properties:
        address:
          type: string
        chainId:
          type: integer
        signature:
          type: string
        message:
          type: string
        isPrimary:
          type: boolean
        label:
          type: string
          nullable: true
      required:
        - address
        - chainId
        - signature
        - message
    WalletConnection:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        chainId:
          type: integer
        chainName:
          type: string
        ensName:
          type: string
          nullable: true
        isPrimary:
          type: boolean
        label:
          type: string
          nullable: true
        verified:
          type: boolean
        connectedAt:
          type: string
          format: date-time
      required:
        - id
        - address
        - chainId
        - chainName
        - isPrimary
        - verified
        - connectedAt
    SyncJobResponse:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
        message:
          type: string
      required:
        - jobId
        - status
        - message
...
    PaginatedCollaborations:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CollaborationSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
        - page
        - limit
    CollaborationSummary:
      type: object
      properties:
        id:
          type: string
        senderName:
          type: string
        senderEmail:
          type: string
          nullable: true
        subject:
          type: string
        status:
          type: string
          enum: [new, in_progress, accepted, rejected, archived]
        budget:
          type: string
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        hasUnreadNotes:
          type: boolean
      required:
        - id
        - senderName
        - subject
        - status
        - createdAt
        - hasUnreadNotes
    CollaborationDetail:
      type: object
      properties:
        id:
          type: string
        senderName:
          type: string
        senderEmail:
          type: string
          nullable: true
        senderWallet:
          type: string
          nullable: true
        subject:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, in_progress, accepted, rejected, archived]
        budget:
          type: string
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        deliverables:
          type: array
          items:
            type: string
        labels:
          type: array
          items:
            type: string
        notes:
          type: array
          items:
            $ref: "#/components/schemas/CollaborationNote"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - senderName
        - subject
        - status
        - deliverables
        - labels
        - notes
        - createdAt
        - updatedAt
    CollaborationNote:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - createdAt
    UpdateCollaborationRequest:
      type: object
      properties:
        status:
          type: string
          enum: [new, in_progress, accepted, rejected, archived]
        note:
          type: string
        labels:
          type: array
          items:
            type: string
        deliverables:
          type: array
          items:
            type: string
        budget:
          type: string
        deadline:
          type: string
          format: date-time
    SubmitCollaborationRequest:
      type: object
      properties:
        creatorHandle:
          type: string
        senderName:
          type: string
        senderEmail:
          type: string
        senderWallet:
          type: string
        subject:
          type: string
        description:
          type: string
        budget:
          type: string
        deadline:
          type: string
          format: date-time
        deliverables:
          type: array
          items:
            type: string
      required:
        - creatorHandle
        - senderName
        - subject
...