# lib/api-zod/src/generated/types/collaborationDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kreatorboard API - Creator Operating System
 * OpenAPI spec version: 0.1.0
 */
import type { CollaborationDetailStatus } from "./collaborationDetailStatus";
import type { CollaborationNote } from "./collaborationNote";

export interface CollaborationDetail {
  id: string;
  senderName: string;
  senderEmail?: string | null;
  senderWallet?: string | null;
  subject: string;
  description?: string | null;
  status: CollaborationDetailStatus;
  budget?: string | null;
  deadline?: Date | null;
  deliverables: string[];
  labels: string[];
  notes: CollaborationNote[];
  createdAt: Date;
  updatedAt: Date;
}