# lib/api-zod/src/generated/types/connectWalletRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kreatorboard API - Creator Operating System
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectWalletRequest {
  address: string;
  chainId: number;
  signature: string;
  message: string;
  isPrimary?: boolean;
  label?: string | null;
}